Move ReplicaGroup selector tests to dedicated ReplicaGroupSelectorTest.java
Because there's a lot of Adaptive Routing specific logic that will be added to the ReplicaGroupSelector tests, move it to a new file to isolate the behaviour better.

Moving this before changing any logic allows the diff in the follow-up commits to be cleaner.

Cleaner diffs, logical test separation

Unit test only change.
